+++ app/Http/Controllers/EventController.php
@@ -4,11 +4,12 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Models\Event;
-use App\Models\EventType;
-use Carbon\Carbon;
+// use App\Models\EventType;
+// use Carbon\Carbon;
 use Illuminate\Support\Str;
 use Intervention\Image\Facades\Image;
 Use Illuminate\Support\Facades\File;
+use App\Rules\noimage;
 
 
 class EventController extends Controller
@@ -24,11 +25,15 @@ class EventController extends Controller
             'event_type_id' => ['required', 'numeric', 'integer', 'exists:event_types,id'],
             'event_name' => ['required', 'string', 'max:255'],
             'event_description' => ['required', 'string'],
-            'event_image' => [ 'image', 'sometimes', 'max:2000'],
             'event_date' => ['required', 'date'],
         ]);
 
         if($request->hasFile('event_image')) {
+
+            $request->validate([
+                'event_image' => ['required', 'image', 'max:2000', new noimage],
+            ]);
+
             //get filename with extension
             $filenameWithExt = $request->file('event_image')->getClientOriginalName();
             //get just file name (using standard php function)
@@ -88,51 +93,56 @@ class EventController extends Controller
             'event_type_id' => ['required', 'numeric', 'integer', 'exists:event_types,id'],
             'event_name' => ['required', 'string', 'max:255'],
             'event_description' => ['required', 'string'],
-            'event_image' => [ 'image', 'sometimes', 'max:2000'],
             'event_date' => ['required', 'date'],
         ]);
 
-        if($request->hasFile('event_image')) {
-            //get filename with extension
-            $filenameWithExt = $request->file('event_image')->getClientOriginalName();
-            //get just file name (using standard php function)
-            $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME);
-            //get just extension
-            $extension = $request->file('event_image')->getClientOriginalExtension();
-            //filename to store(uses a time php function to get current time)
-            //this string is a unique name so that file with duplicate name do not get uploaded and
-            //cause problems when viewing(same problem that occured in CISV photo gallery)
-            $filenameToStore= $filename.'_'.time().'.'.Str::lower($extension);
-
-            // Make Folder if it doesn't exist
-            $path = public_path('storage/event_images');
-            if(!File::isDirectory($path)){
-                    File::makeDirectory($path, 0777, true, true);
-            }
-
-            // Resize image if needed and store it in $image variable
-            // Save image to designated folder inside storage
-            if($request->resize_image == 1){
-                $image = Image::make($request->file('event_image'))->resize(2000, 1000);
-            }
-            else{
-                $image = Image::make($request->file('event_image'));
-            }
-
-            $image->save(public_path('storage/event_images/'. $filenameToStore));
-        }
-        else{
-            $filenameToStore = 'noimage.jpg';
-        }
-
         $event = Event::find($id);
         if ($event) {
+
+            if($request->hasFile('event_image')) {
+
+                $request->validate([
+                    'event_image' => ['required', 'image', 'max:2000', new noimage],
+                ]);
+
+                if($event->event_image != "noimage.jpg")
+                    File::delete(public_path('storage/event_images/'.$event->event_image));  //deletes previous file
+
+                //get filename with extension
+                $filenameWithExt = $request->file('event_image')->getClientOriginalName();
+                //get just file name (using standard php function)
+                $filename = pathinfo($filenameWithExt, PATHINFO_FILENAME);
+                //get just extension
+                $extension = $request->file('event_image')->getClientOriginalExtension();
+                //filename to store(uses a time php function to get current time)
+                //this string is a unique name so that file with duplicate name do not get uploaded and
+                //cause problems when viewing(same problem that occured in CISV photo gallery)
+                $filenameToStore= $filename.'_'.time().'.'.Str::lower($extension);
+
+                // Make Folder if it doesn't exist
+                $path = public_path('storage/event_images');
+                if(!File::isDirectory($path)){
+                        File::makeDirectory($path, 0777, true, true);
+                }
+
+                // Resize image if needed and store it in $image variable
+                // Save image to designated folder inside storage
+                if($request->resize_image == 1){
+                    $image = Image::make($request->file('event_image'))->resize(2000, 1000);
+                }
+                else{
+                    $image = Image::make($request->file('event_image'));
+                }
+
+                $image->save(public_path('storage/event_images/'. $filenameToStore));
+            }
+            else{
+                $filenameToStore = 'noimage.jpg';
+            }
+
             $event->event_type_id = $request->event_type_id;
             $event->event_name = $request->event_name;
             $event->event_description = $request->event_description;
-            if($request->hasFile('event_image')){     //works if there is a new image uploaded
-                File::delete(public_path('storage/event_images/'.$event->event_image));  //deletes previous file
-            }
             $event->event_image = $filenameToStore;
             $event->event_date = $request->event_date;
             $event->save();
@@ -146,7 +156,8 @@ class EventController extends Controller
     {        
         $event = Event::find($id);
         if ($event) {
-            File::delete(public_path('storage/event_images/'.$event->event_image));  //deletes file
+            if($event->event_image != "noimage.jpg")
+                File::delete(public_path('storage/event_images/'.$event->event_image));  //deletes file
             $event->delete();
             return response()->json('Event Deleted Successfully !', 201);
         }
